moved the files  from components folder to pages except navbar and footer
