+++ src/index.js
@@ -19,9 +19,9 @@ import Resources from "./pages/resources";
 import UserProfile from "./pages/userprofile";
 import News from "./pages/news";
 
-import AddPosts from "./components/addPosts";
-import AddQuestions from "./components/addQuestions";
-import ShowQuestion from "./components/showQuestion";
+import AddPosts from "./pages/addPosts";
+import AddQuestions from "./pages/addQuestions";
+import ShowQuestion from "./pages/showQuestion";
 
 import "./css/main.css";
 import "./css/about.css";
