+++ src/pages/addPosts.js
